- Deleted the unused BSTAuthenticationToken - Made the BaseAuthenticationToken abstract - Changed BasicAuthenticationHandler to use an STSAuthenticationToken
