+++ platform/security/handler/security-handler-api/src/main/java/org/codice/ddf/security/handler/api/BaseAuthenticationToken.java
@@ -21,7 +21,7 @@ import org.apache.shiro.authc.AuthenticationToken;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-public class BaseAuthenticationToken implements AuthenticationToken {
+public abstract class BaseAuthenticationToken implements AuthenticationToken {
 
   private X509Certificate[] x509Certs;
 
@@ -109,9 +109,7 @@ public class BaseAuthenticationToken implements AuthenticationToken {
     this.requestURI = requestURI;
   }
 
-  public String getCredentialsAsString() {
-    return credentials.toString();
-  }
+  public abstract String getCredentialsAsString();
 
   // IPv6 addresses should be contained within brackets to conform
   // to the spec IETF RFC 2732
