+++ platform/security/handler/security-handler-basic/src/main/java/org/codice/ddf/security/handler/basic/BasicAuthenticationHandler.java
@@ -24,8 +24,8 @@ import javax.ws.rs.core.HttpHeaders;
 import org.apache.commons.lang.StringUtils;
 import org.codice.ddf.platform.filter.FilterChain;
 import org.codice.ddf.security.handler.api.AuthenticationHandler;
-import org.codice.ddf.security.handler.api.BaseAuthenticationToken;
 import org.codice.ddf.security.handler.api.HandlerResult;
+import org.codice.ddf.security.handler.api.STSAuthenticationToken;
 import org.codice.ddf.security.handler.api.STSAuthenticationTokenFactory;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -78,7 +78,7 @@ public class BasicAuthenticationHandler implements AuthenticationHandler {
 
     LOGGER.debug("Doing authentication and authorization for path {}", path);
 
-    BaseAuthenticationToken token = extractAuthenticationInfo(httpRequest);
+    STSAuthenticationToken token = extractAuthenticationInfo(httpRequest);
 
     // we found credentials, attach to result and return with completed status
     if (token != null) {
@@ -119,7 +119,7 @@ public class BasicAuthenticationHandler implements AuthenticationHandler {
     }
   }
 
-  protected BaseAuthenticationToken extractAuthenticationInfo(HttpServletRequest request) {
+  protected STSAuthenticationToken extractAuthenticationInfo(HttpServletRequest request) {
     String authHeader = request.getHeader(HttpHeaders.AUTHORIZATION);
     if (StringUtils.isEmpty(authHeader)) {
       return null;
@@ -132,11 +132,11 @@ public class BasicAuthenticationHandler implements AuthenticationHandler {
    * Extract the Authorization header and parse into a username/password token.
    *
    * @param authHeader the authHeader string from the HTTP request
-   * @return the initialized BaseAuthenticationToken for this username and password combination (or
+   * @return the initialized STSAuthenticationToken for this username and password combination (or
    *     null)
    */
-  protected BaseAuthenticationToken extractAuthInfo(String authHeader, String ip) {
-    BaseAuthenticationToken token = null;
+  protected STSAuthenticationToken extractAuthInfo(String authHeader, String ip) {
+    STSAuthenticationToken token = null;
     authHeader = authHeader.trim();
     String[] parts = authHeader.split(" ");
     if (parts.length == 2) {
