+++ platform/security/sts/security-sts-bstdelegationhandler/src/test/java/org/codice/ddf/security/delegation/BSTDelegationHandlerTest.java
@@ -13,6 +13,7 @@
  */
 package org.codice.ddf.security.delegation;
 
+import static org.codice.ddf.security.delegation.BSTDelegationHandler.BST_VALUE_TYPE;
 import static org.junit.Assert.assertEquals;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
@@ -23,7 +24,6 @@ import org.apache.cxf.sts.token.delegation.TokenDelegationResponse;
 import org.apache.cxf.ws.security.sts.provider.model.secext.BinarySecurityTokenType;
 import org.apache.cxf.ws.security.sts.provider.model.secext.UsernameTokenType;
 import org.apache.wss4j.dom.WSConstants;
-import org.codice.ddf.security.handler.api.BSTAuthenticationToken;
 import org.junit.Test;
 
 public class BSTDelegationHandlerTest {
@@ -31,8 +31,7 @@ public class BSTDelegationHandlerTest {
   public void testCanHandle() {
     BinarySecurityTokenType binarySecurityTokenType = new BinarySecurityTokenType();
     binarySecurityTokenType.setEncodingType(WSConstants.SOAPMESSAGE_NS + "#Base64Binary");
-    binarySecurityTokenType.setValueType(
-        BSTAuthenticationToken.BST_NS + "#" + BSTAuthenticationToken.BST_LN);
+    binarySecurityTokenType.setValueType(BST_VALUE_TYPE);
     ReceivedToken receivedToken = mock(ReceivedToken.class);
     when(receivedToken.getToken()).thenReturn(binarySecurityTokenType);
     BSTDelegationHandler bstDelegationHandler = new BSTDelegationHandler();
@@ -44,8 +43,7 @@ public class BSTDelegationHandlerTest {
   public void testCanNotHandle() {
     BinarySecurityTokenType binarySecurityTokenType = new BinarySecurityTokenType();
     binarySecurityTokenType.setEncodingType(WSConstants.SOAPMESSAGE_NS + "#WrongType");
-    binarySecurityTokenType.setValueType(
-        BSTAuthenticationToken.BST_NS + "#" + BSTAuthenticationToken.BST_LN);
+    binarySecurityTokenType.setValueType(BST_VALUE_TYPE);
     ReceivedToken receivedToken = mock(ReceivedToken.class);
     when(receivedToken.getToken()).thenReturn(binarySecurityTokenType);
     BSTDelegationHandler bstDelegationHandler = new BSTDelegationHandler();
@@ -57,8 +55,7 @@ public class BSTDelegationHandlerTest {
   public void testDelegationAllowed() {
     BinarySecurityTokenType binarySecurityTokenType = new BinarySecurityTokenType();
     binarySecurityTokenType.setEncodingType(WSConstants.SOAPMESSAGE_NS + "#Base64Binary");
-    binarySecurityTokenType.setValueType(
-        BSTAuthenticationToken.BST_NS + "#" + BSTAuthenticationToken.BST_LN);
+    binarySecurityTokenType.setValueType(BST_VALUE_TYPE);
     ReceivedToken receivedToken = mock(ReceivedToken.class);
     when(receivedToken.getToken()).thenReturn(binarySecurityTokenType);
     TokenDelegationParameters tokenDelegationParameters = mock(TokenDelegationParameters.class);
